[refactor] Member 엔티티 email 필드 index 처리
